+++ XMonad/Layout/Fullscreen.hs
@@ -33,7 +33,7 @@ module XMonad.Layout.Fullscreen
 
 import           XMonad
 import           XMonad.Layout.LayoutModifier
-import           XMonad.Layout.NoBorders        (SmartBorder(), smartBorders)
+import           XMonad.Layout.NoBorders        (SmartBorder, smartBorders)
 import           XMonad.Hooks.EwmhDesktops      (fullscreenStartup)
 import           XMonad.Hooks.ManageHelpers     (isFullscreen)
 import           XMonad.Util.WindowProperties
@@ -84,7 +84,8 @@ fullscreenSupport c = c {
     startupHook = startupHook c <+> fullscreenStartup
   }
 
--- | fullscreenSupport with smartBorders support so the border doesn't show when the window is fullscreen
+-- | fullscreenSupport with smartBorders support so the border doesn't
+-- show when the window is fullscreen
 --
 -- > main = xmonad
 -- >      $ fullscreenSupportBorder
@@ -92,7 +93,11 @@ fullscreenSupport c = c {
 fullscreenSupportBorder :: LayoutClass l Window =>
     XConfig l -> XConfig (ModifiedLayout FullscreenFull
     (ModifiedLayout SmartBorder (ModifiedLayout FullscreenFull l)))
-fullscreenSupportBorder c = fullscreenSupport c { layoutHook = smartBorders $ fullscreenFull $ layoutHook c }
+fullscreenSupportBorder c =
+    fullscreenSupport c { layoutHook = smartBorders
+                                       $ fullscreenFull
+                                       $ layoutHook c
+                        }
 
 -- | Messages that control the fullscreen state of the window.
 -- AddFullscreen and RemoveFullscreen are sent to all layouts